Issue #2, updated readme with extensions list, deleted duplicate, fixed typo
