+++ app/src/main/java/co/temy/android/ktx/StringExtensions.kt
@@ -1,6 +1,7 @@
 package co.temy.android.ktx
 
 import android.graphics.Color
+import android.graphics.Typeface
 import android.text.Html
 import android.text.Spannable
 import android.text.SpannableString
@@ -21,7 +22,12 @@ fun String.highlightTerm(term: String): SpannableString {
     val ranges = matches.map { it.range }
     val spannable = SpannableString(this)
     ranges.forEach {
-        spannable.setSpan(ForegroundColorSpan(Color.BLACK), it.start, it.endInclusive + 1, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
+        spannable.setSpan(
+            ForegroundColorSpan(Color.BLACK),
+            it.start,
+            it.endInclusive + 1,
+            Spannable.SPAN_EXCLUSIVE_EXCLUSIVE
+        )
     }
 
     return spannable
@@ -43,13 +49,37 @@ fun String.replaceItalicWithBold(): SpannableString {
     return spannable
 }
 
-fun String.makeTextClickable(clickableText: String, clickableTextColor: Int, clickListener: () -> (Unit)): SpannableString {
+fun Array<out StyleSpan>.replaceItalicInSpannable(spannable: SpannableString) = forEach {
+    if (it.style == Typeface.ITALIC) {
+        spannable.setSpan(
+            StyleSpan(Typeface.BOLD),
+            spannable.getSpanStart(it),
+            spannable.getSpanEnd(it),
+            Spannable.SPAN_EXCLUSIVE_EXCLUSIVE
+        )
+        spannable.setSpan(
+            ForegroundColorSpan(Color.BLACK),
+            spannable.getSpanStart(it),
+            spannable.getSpanEnd(it),
+            Spannable.SPAN_EXCLUSIVE_EXCLUSIVE
+        )
+        spannable.removeSpan(it)
+    }
+}
+
+fun CharSequence.makeTextClickable(
+    clickableText: String,
+    clickableTextColor: Int,
+    clickListener: () -> (Unit)
+): SpannableString {
     val regex = Regex(clickableText)
     val matches = regex.findAll(this)
     val ranges = matches.map { it.range }
     val spannable = SpannableString(this)
     val clickableSpan = object : ClickableSpan() {
-        override fun onClick(widget: View) { clickListener.invoke() }
+        override fun onClick(widget: View) {
+            clickListener.invoke()
+        }
 
         override fun updateDrawState(ds: TextPaint) {
             ds.color = clickableTextColor
